refactor: dont save images to disk
